Testing

+++ file_io/0-read_textfile.c
@@ -13,7 +13,7 @@
  */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-	int o;
+	ssize_t o;
 	ssize_t n;
 	char *buffer;
